+++ poetry/semver/constraints/multi_constraint.py
@@ -36,6 +36,6 @@ class MultiConstraint(BaseConstraint):
         for constraint in self._constraints:
             constraints.append(str(constraint))
 
-        return '[{}]'.format(
+        return '{}'.format(
             (' ' if self._conjunctive else ' || ').join(constraints)
         )