+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Dependencies of each package is now stored in `poetry.lock`.
 - Improved TOML file management.
+- Dependency resolver now respects the root package python version requirements.
 
 ### Fixed